+++ src/main.rs
@@ -23,7 +23,7 @@ use std::path::{self, PathBuf};
 use storages::Storages;
 
 use crate::cmd_args::{BackupSubCommands, Cli, Commands, StorageCommands};
-use devices::{Device, DEVICESFILE, *};
+use devices::{DEVICESFILE, Device};
 
 mod backups;
 mod cmd_args;
@@ -94,9 +94,10 @@ fn main() -> Result<()> {
         }
         Commands::Sync {
             remote_name,
+            use_libgit2,
             use_sshagent,
             ssh_key,
-        } => cmd_sync::cmd_sync(&config_dir, remote_name, use_sshagent, ssh_key)?,
+        } => cmd_sync::cmd_sync(&config_dir, remote_name, use_sshagent, ssh_key, use_libgit2)?,
         Commands::Status {
             path,
             storage,