+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 # Changelog
 
+## [Unreleased]
+
 ## [0.4.0] - 2025-03-01
 
 ### Added
@@ -49,6 +51,7 @@
 - `backup done` subcommand
 - `completion` subcommand
 
+[Unreleased]: https://github.com/qwjyh/xdbm/compare/v0.4.0...HEAD
 [0.4.0]: https://github.com/qwjyh/xdbm/compare/v0.3.0...v0.4.0
 [0.3.0]: https://github.com/qwjyh/xdbm/compare/v0.2.1...v0.3.0
 [0.2.1]: https://github.com/qwjyh/xdbm/compare/v0.2.0...v0.2.1